Move vnode types to header
